requirements.txt
langchain
langchain-huggingface
python-dotenv